refactor(cli): extract roles/projects/hooks/tunnels/handoff/mcp to *_cli.py (#495) (#534)
Phase 1 (Wave 2) of #495. Extracts integrations/config domains into
roles_cli.py / hooks_cli.py / tunnels_cli.py / handoff_cli.py / mcp_cli.py
(hook+skill installer helpers move with hooks). doctor_cli's deferred reads of
the hook/skill helpers repointed to hooks_cli. Pure relocation, CLI surface
unchanged. Part of #495.

Built by dotdev.dev
