+++ agentwire/__main__.py
@@ -1,10 +1,7 @@
 """CLI entry point for AgentWire."""
 
 import argparse
-import datetime
-import importlib.resources
 import json
-import re
 import shutil
 import subprocess
 import sys
@@ -74,8 +71,6 @@ from .core import (  # noqa: E402,F401  # E402: must follow load_dotenv(); F401:
     wait_for_shell_prompt,
 )
 from .project_config import (  # noqa: E402  # must follow load_dotenv() above
-    ProjectConfig,
-    SessionType,
     detect_default_agent_type,
     get_parent_from_config,
 )
@@ -903,1394 +898,6 @@ def cmd_uninstall(args) -> int:
     return 0
 
 
-# === MCP Server Command ===
-
-
-def cmd_mcp(args) -> int:
-    """Run the MCP server on stdio.
-
-    Exposes AgentWire capabilities as MCP tools for external agents
-    like MoltBot, Claude Desktop, etc.
-    """
-    from .mcp_server import run_server
-    run_server()
-    return 0
-
-
-# =============================================================================
-# Handoff Commands (shareable conversation bundles — ai-handoff.md + show-the-story.html)
-# =============================================================================
-
-
-_HANDOFF_TEMPLATE = """\
-<session_bundle version="1">
-<title>{{ short title — what this session is about }}</title>
-
-<metadata>
-- cwd: __CWD__
-- repo_url: __REPO_URL__
-- branch: __BRANCH__
-- commit: __COMMIT__
-- session_type: claude-bypass
-- model: __MODEL__
-- started_at: {{ ISO timestamp }}
-- ended_at: {{ ISO timestamp }}
-- user_identity: {{ who }}
-- mcp_servers: {{ comma-separated list }}
-</metadata>
-
-<environment>
-{{ panes, channels, scheduler state, anything else the receiver can't see from cwd alone }}
-</environment>
-
-<instructions>
-{{ The CLI prefilled this section with collected CLAUDE.md / rules / memory.
-   Leave it alone unless you need to redact secrets. }}
-__INSTRUCTIONS_BLOCK__
-</instructions>
-
-<project_state>
-<git_status>
-__GIT_STATUS__
-</git_status>
-<git_log>
-__GIT_LOG__
-</git_log>
-<git_diff>
-__GIT_DIFF__
-</git_diff>
-{{ optionally <file path="..."> blocks for key files }}
-</project_state>
-
-<conversation_summary>
-<goal>{{ one sentence: what this session set out to do }}</goal>
-<tldr>{{ one paragraph the receiver reads first }}</tldr>
-<decisions>
-<decision>
-<title>{{ short name of decision }}</title>
-<rationale>{{ why it was made }}</rationale>
-<alternatives>
-- {{ alternative 1 }}
-- {{ alternative 2 }}
-</alternatives>
-</decision>
-</decisions>
-<dead_ends>
-<dead_end>
-<title>{{ thing tried }}</title>
-<why>{{ why rejected — saves the receiver from retracing }}</why>
-</dead_end>
-</dead_ends>
-<open_threads>
-<thread>
-<title>{{ unresolved item }}</title>
-<note>{{ where to pick up }}</note>
-</thread>
-</open_threads>
-<stats>
-- turns: {{ N }}
-- files_touched: {{ N }}
-- tools_used: {{ N }}
-- duration_minutes: {{ N }}
-</stats>
-</conversation_summary>
-
-<journey>
-<beat title="{{ short name }}">
-<quote>{{ optional verbatim line from the conversation }}</quote>
-<what_happened>{{ what changed at this beat }}</what_happened>
-</beat>
-</journey>
-
-<recent_turns>
-{{ Last ~10-20 turns, filtered: drop tool noise, keep user turns + decision-making
-   assistant turns. Use markdown blockquotes or simple labelled blocks. }}
-</recent_turns>
-
-<handoff>
-<one_sentence>{{ what the next agent should do first }}</one_sentence>
-<resume_at>{{ specific file path / TODO / step number }}</resume_at>
-<caveats>
-- {{ permission boundary, e.g. "do not push" }}
-- {{ env-specific note }}
-</caveats>
-</handoff>
-
-<theme>
-{
-  "name": "{{ short slug, e.g. evening-debug }}",
-  "mood": "{{ honest read of the session's emotional tone }}",
-  "palette": {
-    "bg": "#0e0f13",
-    "surface": "#1a1d24",
-    "fg": "#e2e8f0",
-    "muted": "#64748b",
-    "accent": "#5eead4",
-    "accent_2": "#fbbf24",
-    "border": "#2a2f3a"
-  },
-  "fonts": {
-    "heading": "ui-monospace, 'JetBrains Mono', monospace",
-    "body": "ui-sans-serif, system-ui, sans-serif"
-  },
-  "motion": "subtle"
-}
-</theme>
-</session_bundle>
-"""
-
-
-def _handoff_artifacts_dir() -> Path:
-    try:
-        from .config import load_config
-        cfg = load_config()
-        return Path(str(cfg.artifacts.dir)).expanduser()
-    except Exception:
-        return Path.home() / ".agentwire" / "artifacts"
-
-
-def _handoff_slug(title_hint: str | None = None) -> str:
-    import re as _re
-    ts = datetime.datetime.now().strftime("%Y-%m-%dT%H-%M-%S")
-    if title_hint:
-        clean = _re.sub(r"[^a-z0-9]+", "-", title_hint.lower()).strip("-")[:40]
-        if clean:
-            return f"handoff-{ts}-{clean}"
-    return f"handoff-{ts}"
-
-
-def _build_template(cwd: Path) -> str:
-    """Pre-fill template with collected git + instructions chain."""
-    from .handoff import git_state, instructions
-
-    snap = git_state.snapshot(cwd)
-    chain = instructions.collect(cwd)
-
-    instructions_block_parts = []
-    for instr in chain:
-        instructions_block_parts.append(
-            f'<file path="{instr.path}" kind="{instr.kind}">\n{instr.content}\n</file>'
-        )
-    instructions_block = "\n".join(instructions_block_parts) if instructions_block_parts else \
-        "{{ no CLAUDE.md or memory files found — this bundle is portable but light on instructions }}"
-
-    substitutions = {
-        "__CWD__": str(cwd),
-        "__REPO_URL__": snap.get("remote_url") or "{{ none }}",
-        "__BRANCH__": snap.get("branch") or "{{ unknown }}",
-        "__COMMIT__": snap.get("commit") or "{{ unknown }}",
-        "__MODEL__": "claude-opus-4-7",
-        "__INSTRUCTIONS_BLOCK__": instructions_block,
-        "__GIT_STATUS__": snap.get("status") or "(clean)",
-        "__GIT_LOG__": snap.get("log") or "(no commits)",
-        "__GIT_DIFF__": snap.get("diff") or "(no uncommitted diff)",
-    }
-    text = _HANDOFF_TEMPLATE
-    for key, value in substitutions.items():
-        text = text.replace(key, value)
-    return text
-
-
-def cmd_handoff_init(args) -> int:
-    """Create a handoff bundle directory and emit a pre-filled ai-handoff.md template.
-
-    The agent then edits ai-handoff.md to add the session-specific summary,
-    decisions, journey, and theme — the parts only the agent has context for.
-    """
-    json_mode = getattr(args, 'json', False)
-    title_hint = getattr(args, 'title', None)
-    output_dir = getattr(args, 'output_dir', None)
-
-    base = Path(output_dir).expanduser() if output_dir else _handoff_artifacts_dir()
-    bundle_dir = base / _handoff_slug(title_hint)
-    bundle_dir.mkdir(parents=True, exist_ok=True)
-
-    cwd = Path.cwd()
-    template_text = _build_template(cwd)
-    handoff_path = bundle_dir / "ai-handoff.md"
-    handoff_path.write_text(template_text, encoding="utf-8")
-
-    manifest = {
-        "schema_version": 1,
-        "created_at": datetime.datetime.now().isoformat(timespec="seconds"),
-        "cwd": str(cwd),
-        "title_hint": title_hint or "",
-    }
-    (bundle_dir / "manifest.json").write_text(
-        json.dumps(manifest, indent=2), encoding="utf-8"
-    )
-
-    if json_mode:
-        _output_json({
-            "success": True,
-            "bundle_dir": str(bundle_dir),
-            "ai_handoff_path": str(handoff_path),
-            "manifest_path": str(bundle_dir / "manifest.json"),
-            "instructions": (
-                "Edit ai-handoff.md to fill in the {{ ... }} placeholders. "
-                "Then run: agentwire handoff render <bundle_dir> --story"
-            ),
-        })
-    else:
-        print(f"Bundle dir: {bundle_dir}")
-        print(f"Edit:       {handoff_path}")
-        print()
-        print("Next: fill in the placeholders, then run:")
-        print(f"  agentwire handoff render {bundle_dir} --story")
-    return 0
-
-
-def cmd_handoff_render(args) -> int:
-    """Render show-the-story.html from an existing ai-handoff.md."""
-    from .handoff.parser import HandoffParseError, parse_file
-    from .handoff.renderer import render_to_file
-
-    json_mode = getattr(args, 'json', False)
-    target = Path(args.path).expanduser()
-
-    if target.is_dir():
-        md_path = target / "ai-handoff.md"
-        bundle_dir = target
-    else:
-        md_path = target
-        bundle_dir = target.parent
-
-    if not md_path.exists():
-        msg = f"ai-handoff.md not found at {md_path}"
-        if json_mode:
-            _output_json({"success": False, "error": msg})
-        else:
-            print(msg, file=sys.stderr)
-        return 1
-
-    try:
-        bundle = parse_file(md_path)
-    except HandoffParseError as e:
-        msg = f"Invalid ai-handoff.md: {e}"
-        if json_mode:
-            _output_json({"success": False, "error": msg, "path": str(md_path)})
-        else:
-            print(msg, file=sys.stderr)
-        return 1
-
-    outputs: dict[str, str] = {"ai_handoff_path": str(md_path)}
-
-    if getattr(args, 'story', True):
-        story_path = bundle_dir / "show-the-story.html"
-        render_to_file(bundle, story_path)
-        outputs["show_the_story_path"] = str(story_path)
-
-    if json_mode:
-        _output_json({"success": True, "bundle_dir": str(bundle_dir), **outputs})
-    else:
-        print(f"Bundle dir: {bundle_dir}")
-        for k, v in outputs.items():
-            print(f"  {k}: {v}")
-    return 0
-
-
-def cmd_handoff_list(args) -> int:
-    """List past handoff bundles in the artifacts directory."""
-    json_mode = getattr(args, 'json', False)
-    base = _handoff_artifacts_dir()
-
-    bundles = []
-    if base.exists():
-        for d in sorted(base.glob("handoff-*")):
-            if not d.is_dir():
-                continue
-            handoff_md = d / "ai-handoff.md"
-            story_html = d / "show-the-story.html"
-            manifest_path = d / "manifest.json"
-            manifest_data = {}
-            if manifest_path.exists():
-                try:
-                    manifest_data = json.loads(manifest_path.read_text(encoding="utf-8"))
-                except json.JSONDecodeError:
-                    pass
-            bundles.append({
-                "name": d.name,
-                "path": str(d),
-                "ai_handoff_exists": handoff_md.exists(),
-                "show_the_story_exists": story_html.exists(),
-                "created_at": manifest_data.get("created_at", ""),
-                "title_hint": manifest_data.get("title_hint", ""),
-                "cwd": manifest_data.get("cwd", ""),
-            })
-
-    if json_mode:
-        _output_json({"success": True, "bundles": bundles, "count": len(bundles)})
-    else:
-        if not bundles:
-            print(f"No handoff bundles in {base}")
-            return 0
-        print(f"Handoff bundles in {base}:")
-        for b in bundles:
-            badge = []
-            if b["ai_handoff_exists"]:
-                badge.append("md")
-            if b["show_the_story_exists"]:
-                badge.append("html")
-            print(f"  {b['name']:<60} [{','.join(badge) or '-'}] {b['title_hint']}")
-    return 0
-
-
-# === Hooks Commands ===
-
-CLAUDE_HOOKS_DIR = Path.home() / ".claude" / "hooks"
-
-
-# =============================================================================
-# Roles Commands
-# =============================================================================
-
-
-def cmd_roles_list(args) -> int:
-    """List available roles from all sources."""
-    from .roles import parse_role_file
-
-    json_mode = getattr(args, 'json', False)
-
-    # Collect roles from all sources
-    roles_data = []
-
-    # User roles (~/.agentwire/roles/)
-    user_roles_dir = Path.home() / ".agentwire" / "roles"
-    if user_roles_dir.exists():
-        for role_file in user_roles_dir.glob("*.md"):
-            role = parse_role_file(role_file)
-            if role:
-                roles_data.append({
-                    "name": role.name,
-                    "description": role.description,
-                    "source": "user",
-                    "path": str(role_file),
-                    "disallowed_tools": role.disallowed_tools,
-                })
-
-    # Bundled roles (agentwire/roles/)
-    try:
-        bundled_dir = Path(__file__).parent / "roles"
-        if bundled_dir.exists():
-            for role_file in bundled_dir.glob("*.md"):
-                # Skip if user already has this role
-                if any(r["name"] == role_file.stem for r in roles_data):
-                    continue
-                role = parse_role_file(role_file)
-                if role:
-                    roles_data.append({
-                        "name": role.name,
-                        "description": role.description,
-                        "source": "bundled",
-                        "path": str(role_file),
-                        "disallowed_tools": role.disallowed_tools,
-                        })
-    except Exception:
-        pass
-
-    if json_mode:
-        _output_json({"roles": roles_data})
-        return 0
-
-    if not roles_data:
-        print("No roles found.")
-        print("Create roles in: ~/.agentwire/roles/")
-        return 0
-
-    # Print table
-    print("Available Roles:")
-    print()
-    print(f"{'Name':<20} {'Source':<10} {'Description':<40}")
-    print("-" * 70)
-    for r in sorted(roles_data, key=lambda x: x["name"]):
-        desc = r["description"][:37] + "..." if len(r["description"]) > 40 else r["description"]
-        print(f"{r['name']:<20} {r['source']:<10} {desc:<40}")
-
-    print()
-    print("User roles: ~/.agentwire/roles/")
-    print("Use 'agentwire roles show <name>' for details")
-    return 0
-
-
-def cmd_projects_list(args) -> int:
-    """List discovered projects."""
-    from .projects import get_projects
-
-    json_mode = getattr(args, 'json', False)
-    machine_filter = getattr(args, 'machine', None)
-
-    projects = get_projects(machine=machine_filter)
-
-    if json_mode:
-        _output_json({"projects": projects})
-        return 0
-
-    if not projects:
-        print("No projects found.")
-        print("Projects need a .agentwire.yml file in their directory.")
-        return 0
-
-    # Print table
-    print(f"Discovered Projects ({len(projects)}):\n")
-    print(f"{'Name':<25} {'Type':<15} {'Path':<40}")
-    print("-" * 80)
-    for p in projects:
-        # Truncate long paths
-        path = p["path"]
-        if len(path) > 40:
-            path = "..." + path[-37:]
-        machine_suffix = f" @{p['machine']}" if p['machine'] != 'local' else ""
-        print(f"{p['name']:<25} {p['type']:<15} {path:<40}{machine_suffix}")
-
-    print()
-    return 0
-
-
-_VALID_PROJECT_NAME = re.compile(r"^[A-Za-z0-9][A-Za-z0-9._-]*$")
-
-
-def cmd_projects_create(args) -> int:
-    """Create a new local project: make the directory, optionally git-init or clone, and write .agentwire.yml."""
-    from .config import get_config
-    from .project_config import ensure_gitignored, save_project_config
-
-    name = (args.name or "").strip()
-    clone_url = (getattr(args, "from_url", None) or "").strip() or None
-    do_git_init = bool(getattr(args, "git_init", False))
-    json_mode = getattr(args, "json", False)
-
-    def _fail(msg: str) -> int:
-        if json_mode:
-            _output_json({"success": False, "error": msg})
-        else:
-            print(f"Error: {msg}", file=sys.stderr)
-        return 1
-
-    if not name:
-        return _fail("Project name is required")
-    if not _VALID_PROJECT_NAME.match(name) or "/" in name or ".." in name:
-        return _fail("Invalid project name (allowed: letters, digits, '.', '_', '-')")
-
-    projects_dir = get_config().projects.dir.expanduser().resolve()
-    projects_dir.mkdir(parents=True, exist_ok=True)
-    project_path = projects_dir / name
-
-    if project_path.exists():
-        return _fail(f"Project already exists at {project_path}")
-
-    try:
-        if clone_url:
-            result = subprocess.run(
-                ["git", "clone", clone_url, str(project_path)],
-                capture_output=True,
-                text=True,
-                timeout=300,
-            )
-            if result.returncode != 0:
-                return _fail(f"git clone failed: {result.stderr.strip() or 'unknown error'}")
-        else:
-            project_path.mkdir(parents=True, exist_ok=False)
-            if do_git_init:
-                result = subprocess.run(
-                    ["git", "init"],
-                    cwd=str(project_path),
-                    capture_output=True,
-                    text=True,
-                    timeout=30,
-                )
-                if result.returncode != 0:
-                    return _fail(f"git init failed: {result.stderr.strip() or 'unknown error'}")
-    except subprocess.TimeoutExpired:
-        return _fail("Operation timed out")
-    except OSError as e:
-        return _fail(f"Failed to create directory: {e}")
-
-    gitignore_updated = ensure_gitignored(project_path)
-    config = ProjectConfig(type=SessionType.from_str("claude-bypass"), roles=[], voice=None)
-    if not save_project_config(config, project_path):
-        return _fail("Created project directory but failed to write .agentwire.yml")
-
-    payload = {
-        "success": True,
-        "name": name,
-        "path": str(project_path),
-        "machine": "local",
-        "cloned": bool(clone_url),
-        "git_init": do_git_init and not clone_url,
-    }
-    if json_mode:
-        _output_json(payload)
-    else:
-        print(f"Created project '{name}' at {project_path}")
-        if clone_url:
-            print(f"  Cloned from: {clone_url}")
-        elif do_git_init:
-            print("  Initialized empty git repository")
-        if gitignore_updated:
-            print("  Added .agentwire.yml to .gitignore (personal config — keep it untracked)")
-    return 0
-
-
-def cmd_roles_show(args) -> int:
-    """Show details for a specific role."""
-    from .roles import discover_role, parse_role_file
-
-    name = args.name
-    json_mode = getattr(args, 'json', False)
-
-    # Discover role
-    role_path = discover_role(name)
-    if not role_path:
-        if json_mode:
-            _output_json({"error": f"Role '{name}' not found"})
-        else:
-            print(f"Role '{name}' not found.", file=sys.stderr)
-            print("Available locations:")
-            print(f"  User: ~/.agentwire/roles/{name}.md")
-            print(f"  Bundled: agentwire/roles/{name}.md")
-        return 1
-
-    role = parse_role_file(role_path)
-    if not role:
-        if json_mode:
-            _output_json({"error": "Failed to parse role file"})
-        else:
-            print(f"Failed to parse role file: {role_path}", file=sys.stderr)
-        return 1
-
-    if json_mode:
-        _output_json({
-            "name": role.name,
-            "description": role.description,
-            "path": str(role_path),
-            "tools": role.tools,
-            "disallowed_tools": role.disallowed_tools,
-            "color": role.color,
-            "instructions": role.instructions,
-        })
-        return 0
-
-    print(f"Role: {role.name}")
-    print(f"Description: {role.description or '(none)'}")
-    print(f"Path: {role_path}")
-    if role.tools:
-        print(f"Tools (whitelist): {', '.join(role.tools)}")
-    if role.disallowed_tools:
-        print(f"Disallowed Tools: {', '.join(role.disallowed_tools)}")
-    print()
-    if role.instructions:
-        print("Instructions:")
-        print("-" * 40)
-        print(role.instructions)
-        print("-" * 40)
-    else:
-        print("Instructions: (none)")
-
-    return 0
-
-
-def get_hooks_source() -> Path:
-    """Get the path to the hooks directory in the installed package."""
-    # First try: hooks directory inside the agentwire package
-    package_dir = Path(__file__).parent
-    hooks_dir = package_dir / "hooks"
-    if hooks_dir.exists():
-        return hooks_dir
-
-    # Fallback: try importlib.resources (for installed packages)
-    try:
-        with importlib.resources.files("agentwire").joinpath("hooks") as p:
-            if p.exists():
-                return Path(p)
-    except (TypeError, FileNotFoundError):
-        pass
-
-    raise FileNotFoundError("Could not find hooks directory in package")
-
-
-CLAUDE_COMMANDS_DIR = Path.home() / ".claude" / "commands"
-
-
-def get_commands_source() -> Path:
-    """Get the path to the commands directory in the installed package."""
-    package_dir = Path(__file__).parent
-    commands_dir = package_dir / "commands"
-    if commands_dir.exists():
-        return commands_dir
-
-    try:
-        with importlib.resources.files("agentwire").joinpath("commands") as p:
-            if p.exists():
-                return Path(p)
-    except (TypeError, FileNotFoundError):
-        pass
-
-    raise FileNotFoundError("Could not find commands directory in package")
-
-
-def install_commands(force: bool = False) -> list[str]:
-    """Symlink bundled slash commands into ~/.claude/commands/.
-
-    Returns list of command names that were installed or updated.
-    """
-    try:
-        commands_source = get_commands_source()
-    except FileNotFoundError:
-        return []
-
-    CLAUDE_COMMANDS_DIR.mkdir(parents=True, exist_ok=True)
-
-    installed = []
-    for src_file in commands_source.glob("*.md"):
-        target = CLAUDE_COMMANDS_DIR / src_file.name
-        if target.exists() or target.is_symlink():
-            if target.is_symlink() and target.resolve() == src_file.resolve() and not force:
-                continue  # Already correctly symlinked
-            target.unlink()
-
-        target.symlink_to(src_file.resolve())
-        installed.append(src_file.stem)
-
-    return installed
-
-
-CLAUDE_SKILLS_DIR = Path.home() / ".claude" / "skills"
-
-
-def get_skills_source() -> Path:
-    """Get the path to the skills directory in the installed package.
-
-    Mirrors get_commands_source(): resolves to the bundled package dir so the
-    installed symlink is auto-current after `agentwire rebuild` (which reinstalls
-    the wheel), never a transient checkout path.
-    """
-    package_dir = Path(__file__).parent
-    skills_dir = package_dir / "skills"
-    if skills_dir.exists():
-        return skills_dir
-
-    try:
-        with importlib.resources.files("agentwire").joinpath("skills") as p:
-            if p.exists():
-                return Path(p)
-    except (TypeError, FileNotFoundError):
-        pass
-
-    raise FileNotFoundError("Could not find skills directory in package")
-
-
-def _managed_global_skills() -> list[str]:
-    """Agentwire-owned skills that belong GLOBALLY in ~/.claude/skills/.
-
-    Only `wiki` is global — the wiki store lives at ~/.agentwire/wiki/ and is
-    usable from any session. The agentwire-* skills stay project-scoped (shipped
-    via the repo's .claude/skills/, discovered per-project) and are NOT installed
-    here. Third-party skills (cua-driver, shadcn-ui, …) are never touched.
-    """
-    return ["wiki"]
-
-
-def _managed_skill_state(target: Path, source: Path) -> str:
-    """Drift state of a managed global skill DIRECTORY: missing | stale | ok.
-
-    Skills are directories, so unlike _managed_file_state this never compares
-    bytes. A symlink is ok only when it resolves to the packaged source; a real
-    directory (the hand-placed pre-#475 state) or a symlink pointing elsewhere is
-    stale and must be removed before re-symlinking.
-    """
-    if target.is_symlink():
-        if not target.exists():
-            return "stale"  # dangling symlink
-        return "ok" if target.resolve() == source.resolve() else "stale"
-    if not target.exists():
-        return "missing"
-    return "stale"  # real dir/file occupying the slot
-
-
-def _remove_skill_target(target: Path) -> None:
-    """Clear whatever occupies a skill slot — symlink, real dir, or stray file."""
-    if target.is_symlink() or target.is_file():
-        target.unlink()
-    elif target.is_dir():
-        shutil.rmtree(target)
-
-
-def install_skills(force: bool = False, copy: bool = False) -> dict[str, str]:
-    """Install/refresh agentwire-owned global skills into ~/.claude/skills/.
-
-    Each managed skill is a directory installed as a symlink (or copied with
-    --copy) pointing at the packaged source, drift-aware: a correct symlink is
-    left alone, a real-dir / wrong-symlink target is replaced.
-
-    Returns {name: "installed" | "updated" | "current" | "missing-source"}.
-    """
-    try:
-        skills_source = get_skills_source()
-    except FileNotFoundError:
-        print("  Warning: skills directory not found, skipping skill installation")
-        return {}
-
-    results: dict[str, str] = {}
-    for name in _managed_global_skills():
-        source = skills_source / name
-        if not source.exists():
-            print(f"  Warning: skill '{name}' not found in package, skipping")
-            results[name] = "missing-source"
-            continue
-
-        target = CLAUDE_SKILLS_DIR / name
-        state = _managed_skill_state(target, source)
-        if state == "ok" and not force:
-            results[name] = "current"
-            continue
-
-        existed = target.exists() or target.is_symlink()
-        CLAUDE_SKILLS_DIR.mkdir(parents=True, exist_ok=True)
-        _remove_skill_target(target)
-        if copy:
-            shutil.copytree(source, target)
-        else:
-            target.symlink_to(source.resolve(), target_is_directory=True)
-        results[name] = "updated" if existed else "installed"
-
-    return results
-
-
-def skill_drift() -> dict[str, str]:
-    """Drift state of agentwire-owned global skills.
-
-    Returns {name: ok|stale|missing|source-unavailable}. Mirrors
-    cli_safety.*_drift() so `agentwire doctor` can flag a hand-placed or drifted
-    skill the same way it flags hook drift.
-
-    `source-unavailable` means the packaged skill can't be resolved in the
-    running context — e.g. doctor invoked from a SOURCE checkout, where skills
-    only exist inside the built wheel (`agentwire/skills/`), not on disk. That is
-    NOT a drift problem: there is nothing to install from, so doctor skips it
-    rather than crying "missing". `missing`/`stale` are reserved for the case
-    where the source IS resolvable (installed tool) and the installed copy is
-    genuinely absent or wrong.
-    """
-    try:
-        skills_source = get_skills_source()
-    except FileNotFoundError:
-        return {name: "source-unavailable" for name in _managed_global_skills()}
-
-    drift: dict[str, str] = {}
-    for name in _managed_global_skills():
-        source = skills_source / name
-        target = CLAUDE_SKILLS_DIR / name
-        if not source.exists():
-            drift[name] = "source-unavailable"
-            continue
-        drift[name] = _managed_skill_state(target, source)
-    return drift
-
-
-def register_hook_in_settings(event: str, hook_name: str) -> bool:
-    """Register a hook under `event` in Claude's settings.json.
-
-    Returns True if settings were updated, False if already configured.
-
-    Claude Code hook format:
-    {
-      "hooks": {
-        "<event>": [
-          {
-            "matcher": ".*",
-            "hooks": [
-              {"type": "command", "command": "~/.claude/hooks/<hook_name>"}
-            ]
-          }
-        ]
-      }
-    }
-    """
-    settings_file = Path.home() / ".claude" / "settings.json"
-    # Use ~ for portability
-    hook_command = f"~/.claude/hooks/{hook_name}"
-
-    # Load existing settings or create new
-    if settings_file.exists():
-        try:
-            settings = json.loads(settings_file.read_text())
-        except json.JSONDecodeError:
-            settings = {}
-    else:
-        settings = {}
-
-    # Ensure hooks structure exists
-    if "hooks" not in settings:
-        settings["hooks"] = {}
-    if event not in settings["hooks"]:
-        settings["hooks"][event] = []
-
-    # Check if already registered (check nested hooks array)
-    for entry in settings["hooks"][event]:
-        if "hooks" in entry:
-            for h in entry["hooks"]:
-                if h.get("command") == hook_command:
-                    return False  # Already registered
-
-    # Add hook with correct Claude Code format
-    hook_entry = {
-        "matcher": ".*",
-        "hooks": [
-            {"type": "command", "command": hook_command}
-        ]
-    }
-    settings["hooks"][event].append(hook_entry)
-
-    # Write back
-    settings_file.parent.mkdir(parents=True, exist_ok=True)
-    settings_file.write_text(json.dumps(settings, indent=2))
-
-    return True
-
-
-# Agentwire-owned files deployed by `hooks install`. Each entry:
-# (filename in agentwire/hooks/, target directory, settings.json event or None)
-def _managed_hook_files() -> list[tuple[str, Path, str | None]]:
-    return [
-        ("agentwire-permission.sh", CLAUDE_HOOKS_DIR, "PermissionRequest"),
-        ("idle-handler.sh", CLAUDE_HOOKS_DIR, "Notification"),
-        ("queue-processor.sh", Path.home() / ".agentwire", None),
-    ]
-
-
-def _managed_file_state(target: Path, source: Path) -> str:
-    """Drift state of an agentwire-managed installed file: missing | stale | ok.
-
-    Symlinks are ok when they resolve to the packaged source; regular files
-    are ok when their content matches it byte-for-byte.
-    """
-    if target.is_symlink():
-        if not target.exists():
-            return "stale"  # dangling symlink
-        return "ok" if target.resolve() == source.resolve() else "stale"
-    if not target.exists():
-        return "missing"
-    try:
-        return "ok" if target.read_bytes() == source.read_bytes() else "stale"
-    except OSError:
-        return "stale"
-
-
-def _install_managed_file(source: Path, target: Path, force: bool = False, copy: bool = False) -> bool:
-    """Install or refresh an agentwire-owned file (symlink by default).
-
-    These files carry no user edits to preserve — any drift from the packaged
-    source is replaced. Returns True if the target was created or updated.
-    """
-    if not force and _managed_file_state(target, source) == "ok":
-        return False
-
-    target.parent.mkdir(parents=True, exist_ok=True)
-    if target.exists() or target.is_symlink():
-        target.unlink()
-    if copy:
-        shutil.copy2(source, target)
-    else:
-        target.symlink_to(source.resolve())
-    target.chmod(0o755)
-    return True
-
-
-def install_hooks(force: bool = False, copy: bool = False) -> dict[str, str]:
-    """Install/refresh all agentwire-owned hook files + settings.json registration.
-
-    Returns {filename: "installed" | "updated" | "current" | "missing-source"}.
-    """
-    try:
-        hooks_source = get_hooks_source()
-    except FileNotFoundError:
-        print("  Warning: hooks directory not found, skipping hook installation")
-        return {}
-
-    results: dict[str, str] = {}
-    for hook_name, target_dir, event in _managed_hook_files():
-        source = hooks_source / hook_name
-        if not source.exists():
-            print(f"  Warning: {hook_name} not found in package, skipping")
-            results[hook_name] = "missing-source"
-            continue
-
-        target = target_dir / hook_name
-        existed = target.exists() or target.is_symlink()
-        if _install_managed_file(source, target, force=force, copy=copy):
-            results[hook_name] = "updated" if existed else "installed"
-        else:
-            results[hook_name] = "current"
-
-        if event:
-            register_hook_in_settings(event, hook_name)
-
-    # Heal the full damage-control surface (hook scripts + rules + tooldefs +
-    # PreToolUse matchers), not just the settings.json matchers. This closes the
-    # documented post-rebuild gap: CLAUDE.md tells users to re-run `hooks install`
-    # after a rebuild, so it must actually sync the DC files/rules — drift-aware,
-    # never clobbering a customized rule.
-    try:
-        from agentwire.cli_safety import heal_damage_control
-        heal_damage_control(quiet=True)
-    except Exception:
-        pass
-
-    # Global skills (currently just /wiki) are agentwire-owned too, and rotted
-    # silently because nothing ever resynced the hand-placed copies. Heal them
-    # on the same install pass — drift-aware, like the hooks above.
-    results.update(install_skills(force=force, copy=copy))
-
-    return results
-
-
-def cmd_hooks_install(args) -> int:
-    """Install agentwire-owned hook files and slash commands for AgentWire integration."""
-    results = install_hooks(force=args.force, copy=args.copy)
-    for hook_name, target_dir, _event in _managed_hook_files():
-        state = results.get(hook_name)
-        if state in ("installed", "updated"):
-            print(f"{state.capitalize()} {hook_name} -> {target_dir / hook_name}")
-        elif state == "current":
-            print(f"{hook_name} already current.")
-
-    installed_commands = install_commands(force=args.force)
-    if installed_commands:
-        print(f"\nInstalled slash commands to {CLAUDE_COMMANDS_DIR}:")
-        for name in installed_commands:
-            print(f"  /{name}")
-    else:
-        print("Slash commands already installed.")
-
-    refreshed_skills = False
-    for name in _managed_global_skills():
-        state = results.get(name)
-        if state in ("installed", "updated"):
-            print(f"{state.capitalize()} skill -> /{name} ({CLAUDE_SKILLS_DIR / name})")
-            refreshed_skills = True
-    if not refreshed_skills:
-        print("Skills already current.")
-
-    return 0
-
-
-def unregister_hook_from_settings(event: str, hook_name: str) -> bool:
-    """Remove a hook registered under `event` from Claude's settings.json.
-
-    Returns True if settings were updated, False if not found.
-    """
-    settings_file = Path.home() / ".claude" / "settings.json"
-    hook_command = f"~/.claude/hooks/{hook_name}"
-
-    if not settings_file.exists():
-        return False
-
-    try:
-        settings = json.loads(settings_file.read_text())
-    except json.JSONDecodeError:
-        return False
-
-    if "hooks" not in settings or event not in settings["hooks"]:
-        return False
-
-    # Filter out entries containing our hook
-    original_len = len(settings["hooks"][event])
-    new_entries = []
-    for entry in settings["hooks"][event]:
-        if "hooks" in entry:
-            # Check if any hook in this entry matches ours
-            has_our_hook = any(h.get("command") == hook_command for h in entry["hooks"])
-            if not has_our_hook:
-                new_entries.append(entry)
-        else:
-            new_entries.append(entry)
-
-    settings["hooks"][event] = new_entries
-
-    if len(settings["hooks"][event]) == original_len:
-        return False  # Hook wasn't registered
-
-    # Clean up empty structures
-    if not settings["hooks"][event]:
-        del settings["hooks"][event]
-    if not settings["hooks"]:
-        del settings["hooks"]
-
-    # Write back
-    settings_file.write_text(json.dumps(settings, indent=2))
-    return True
-
-
-def is_hook_registered(event: str, hook_name: str) -> bool:
-    """Check if a hook is registered under `event` in Claude's settings.json."""
-    settings_file = Path.home() / ".claude" / "settings.json"
-    hook_command = f"~/.claude/hooks/{hook_name}"
-
-    if not settings_file.exists():
-        return False
-
-    try:
-        settings = json.loads(settings_file.read_text())
-    except json.JSONDecodeError:
-        return False
-
-    if "hooks" not in settings or event not in settings["hooks"]:
-        return False
-
-    # Check nested hooks array for our command
-    for entry in settings["hooks"][event]:
-        if "hooks" in entry:
-            for h in entry["hooks"]:
-                if h.get("command") == hook_command:
-                    return True
-    return False
-
-
-def cmd_hooks_uninstall(args) -> int:
-    """Remove all agentwire-owned hook files and their settings.json registration."""
-    removed_any = False
-    for hook_name, target_dir, event in _managed_hook_files():
-        target = target_dir / hook_name
-        if target.exists() or target.is_symlink():
-            target.unlink()
-            print(f"Removed: {target}")
-            removed_any = True
-        if event and unregister_hook_from_settings(event, hook_name):
-            print(f"Unregistered {hook_name} from Claude settings.json")
-
-    if not removed_any:
-        print("No hooks installed")
-
-    return 0
-
-
-def cmd_hooks_status(args) -> int:
-    """Check agentwire-owned hook files and tmux portal sync hooks."""
-    print("=== AgentWire Hooks ===")
-    try:
-        hooks_source = get_hooks_source()
-    except FileNotFoundError:
-        hooks_source = None
-
-    for hook_name, target_dir, event in _managed_hook_files():
-        target = target_dir / hook_name
-        print(f"{hook_name}:")
-
-        if not (target.exists() or target.is_symlink()):
-            print("  Status: not installed — run 'agentwire hooks install'")
-            continue
-
-        kind = "symlink" if target.is_symlink() else "copy"
-        if hooks_source and (hooks_source / hook_name).exists():
-            state = _managed_file_state(target, hooks_source / hook_name)
-            drift = "" if state == "ok" else " — STALE, run 'agentwire hooks install'"
-        else:
-            drift = " — packaged source not found, drift unknown"
-        print(f"  Status: installed ({kind}){drift}")
-        location = f"{target} -> {target.resolve()}" if target.is_symlink() else str(target)
-        print(f"  Location: {location}")
-
-        if event:
-            if is_hook_registered(event, hook_name):
-                print(f"  Registered: yes ({event} in ~/.claude/settings.json)")
-            else:
-                print("  Registered: NO - run 'agentwire hooks install' to fix")
-
-    # Tmux portal sync hooks
-    print("\n=== Tmux Portal Sync Hooks ===")
-    try:
-        # Check global hooks first
-        global_result = subprocess.run(
-            ["tmux", "show-hooks", "-g"],
-            capture_output=True,
-            text=True,
-        )
-        global_hooks = global_result.stdout.strip()
-
-        print("Global hooks:")
-        has_global_created = "session-created" in global_hooks
-        has_global_closed = "session-closed" in global_hooks
-
-        if has_global_created or has_global_closed:
-            parts = []
-            if has_global_created:
-                parts.append("session-created")
-            if has_global_closed:
-                parts.append("session-closed")
-            print(f"  {', '.join(parts)}")
-        else:
-            print("  none (run 'agentwire portal restart' to install)")
-
-        # Get list of sessions for per-session hooks
-        result = subprocess.run(
-            ["tmux", "list-sessions", "-F", "#{session_name}"],
-            capture_output=True,
-            text=True,
-        )
-        if result.returncode != 0:
-            print("\nNo tmux sessions running")
-            return 0
-
-        sessions = result.stdout.strip().split("\n") if result.stdout.strip() else []
-
-        if sessions:
-            print("\nPer-session hooks:")
-            for session in sessions:
-                hooks_result = subprocess.run(
-                    ["tmux", "show-hooks", "-t", session],
-                    capture_output=True,
-                    text=True,
-                )
-                hooks_output = hooks_result.stdout.strip()
-
-                has_session_closed = "session-closed" in hooks_output
-                has_kill_pane = "after-kill-pane" in hooks_output
-
-                status_parts = []
-                if has_session_closed:
-                    status_parts.append("session-closed")
-                if has_kill_pane:
-                    status_parts.append("after-kill-pane")
-
-                if status_parts:
-                    print(f"  {session}: {', '.join(status_parts)}")
-                else:
-                    print(f"  {session}: none")
-
-    except Exception as e:
-        print(f"Error checking tmux hooks: {e}")
-
-    return 0
-
-
-# === Tunnel Commands ===
-
-
-def cmd_tunnels_up(args) -> int:
-    """Create all required tunnels."""
-    from .network import NetworkContext
-    from .tunnels import TunnelManager
-
-    ctx = NetworkContext.from_config()
-    manager = TunnelManager()
-    required = ctx.get_required_tunnels()
-
-    if not required:
-        print("No tunnels required for this machine's configuration.")
-        print("(All services run locally or no remote services configured)")
-        return 0
-
-    print("Creating tunnels for this machine...\n")
-
-    all_success = True
-    for i, spec in enumerate(required, 1):
-        # Get service name for display
-        service_name = _get_service_for_tunnel(ctx, spec)
-
-        print(f"[{i}/{len(required)}] {service_name} (localhost:{spec.local_port} -> {spec.remote_machine}:{spec.remote_port})")
-
-        status = manager.create_tunnel(spec, ctx)
-
-        if status.status == "up":
-            if status.error:
-                # Tunnel up but service not responding
-                print(f"      ! Tunnel created (PID {status.pid})")
-                print(f"      ! Warning: {status.error}")
-            else:
-                print(f"      + Tunnel created (PID {status.pid})")
-        else:
-            all_success = False
-            print(f"      x Failed: {status.error}")
-            _print_tunnel_help(spec, status.error)
-
-        print()
-
-    if all_success:
-        print("All tunnels up. Services should be reachable.")
-    else:
-        print("Some tunnels failed. Check errors above.")
-        return 1
-
-    return 0
-
-
-def cmd_tunnels_down(args) -> int:
-    """Tear down all tunnels."""
-    from .tunnels import TunnelManager
-
-    manager = TunnelManager()
-    count = manager.destroy_all_tunnels()
-
-    if count == 0:
-        print("No active tunnels to tear down.")
-    else:
-        print(f"Killed {count} tunnel(s).")
-
-    return 0
-
-
-def cmd_tunnels_status(args) -> int:
-    """Show tunnel health."""
-    from .network import NetworkContext
-    from .tunnels import TunnelManager
-
-    ctx = NetworkContext.from_config()
-    manager = TunnelManager()
-
-    # Get both required and active tunnels
-    required = ctx.get_required_tunnels()
-    active = manager.list_tunnels()
-
-    print("AgentWire Tunnels")
-    print("-" * 55)
-
-    if not required and not active:
-        print("\nNo tunnels configured or active.")
-        print("(All services run locally or no remote services configured)")
-        return 0
-
-    # Show required tunnels
-    for spec in required:
-        service_name = _get_service_for_tunnel(ctx, spec)
-
-        print(f"\n{service_name} (localhost:{spec.local_port} -> {spec.remote_machine}:{spec.remote_port})")
-
-        status = manager.check_tunnel(spec)
-
-        if status.status == "up":
-            print(f"  Status: + UP (PID {status.pid})")
-        elif status.status == "down":
-            print("  Status: - DOWN")
-        else:
-            print("  Status: x ERROR")
-            if status.error:
-                print(f"  Error: {status.error}")
-
-    # Show any orphaned tunnels (active but not required)
-    required_ids = {s.id for s in required}
-    orphaned = [t for t in active if t.spec.id not in required_ids]
-    if orphaned:
-        print("\n" + "-" * 55)
-        print("\nOrphaned tunnels (active but no longer required):")
-        for t in orphaned:
-            print(f"  localhost:{t.spec.local_port} -> {t.spec.remote_machine}:{t.spec.remote_port}")
-            print(f"    PID: {t.pid}, Status: {t.status}")
-
-    print("\n" + "-" * 55)
-
-    # Show next steps
-    down_tunnels = [s for s in required if manager.check_tunnel(s).status != "up"]
-    if down_tunnels:
-        print("To create missing tunnels: agentwire tunnels up")
-
-    return 0
-
-
-def cmd_tunnels_check(args) -> int:
-    """Verify tunnels are working with health checks."""
-    from .network import NetworkContext
-    from .tunnels import TunnelManager, test_service_health
-
-    ctx = NetworkContext.from_config()
-    manager = TunnelManager()
-    required = ctx.get_required_tunnels()
-
-    if not required:
-        print("No tunnels required for this machine.")
-        return 0
-
-    print("Checking tunnel health...\n")
-
-    all_healthy = True
-    for spec in required:
-        service_name = _get_service_for_tunnel(ctx, spec)
-        status = manager.check_tunnel(spec)
-
-        if status.status == "up":
-            # Also test the actual service through the tunnel
-            url = f"http://localhost:{spec.local_port}/health"
-            healthy, err = test_service_health(url, timeout=3)
-
-            if healthy:
-                print(f"+ {service_name}: healthy")
-            else:
-                print(f"! {service_name}: tunnel up but service not responding")
-                if err:
-                    print(f"  {err}")
-                all_healthy = False
-        elif status.status == "down":
-            print(f"x {service_name}: down")
-            all_healthy = False
-        else:
-            print(f"x {service_name}: error - {status.error}")
-            all_healthy = False
-
-    if all_healthy:
-        print("\nAll tunnels healthy.")
-        return 0
-    else:
-        print("\nSome tunnels need attention. Run: agentwire tunnels up")
-        return 1
-
-
-def _get_service_for_tunnel(ctx, spec) -> str:
-    """Get human-readable service name for a tunnel spec."""
-    # Check which service this tunnel is for
-    for service_name in ["portal", "tts"]:
-        service_config = getattr(ctx.config.services, service_name, None)
-        if service_config and service_config.machine == spec.remote_machine and service_config.port == spec.remote_port:
-            return f"Portal -> {service_name.upper()}" if service_name != "portal" else "Portal"
-
-    return f"Tunnel to {spec.remote_machine}"
-
-
-def _print_tunnel_help(spec, error: str) -> None:
-    """Print helpful diagnostics for tunnel errors."""
-    if not error:
-        return
-
-    error_lower = error.lower()
-
-    print("\n      Possible causes:")
-
-    if "port" in error_lower and "in use" in error_lower:
-        print("        1. Another process is using this port")
-        print("        2. A previous tunnel wasn't cleaned up")
-        print("\n      To diagnose:")
-        print(f"        lsof -i :{spec.local_port}    # Find process using port")
-        print("        agentwire tunnels down        # Clean up stale tunnels")
-
-    elif "permission denied" in error_lower:
-        print("        1. SSH key not authorized on remote machine")
-        print("        2. Wrong user configured")
-        print("\n      To fix:")
-        print(f"        ssh-copy-id {spec.remote_machine}")
-
-    elif "host key" in error_lower:
-        print("        1. Remote machine was reinstalled/changed")
-        print("        2. Possible security issue (man-in-the-middle)")
-        print("\n      If expected, fix with:")
-        print(f"        ssh-keygen -R {spec.remote_machine}")
-
-    elif "connection refused" in error_lower:
-        print("        1. SSH server not running on remote")
-        print("        2. Firewall blocking port 22")
-        print("\n      To diagnose:")
-        print(f"        ssh {spec.remote_machine} echo ok")
-
-    elif "timed out" in error_lower or "no route" in error_lower:
-        print("        1. Machine is powered off or unreachable")
-        print("        2. Network connectivity issue")
-        print("\n      To diagnose:")
-        print(f"        ping {spec.remote_machine}")
-
-    elif "not responding" in error_lower:
-        print("        1. Remote service not started")
-        print("        2. Remote service on wrong port")
-        print("\n      To diagnose:")
-        print(f"        ssh {spec.remote_machine} 'lsof -i :{spec.remote_port}'")
-
-
 class VersionAction(argparse.Action):
     """Custom version action that checks Python version and pip environment."""
 
@@ -2385,26 +992,6 @@ Command Categories:
     quo_parser.add_argument("--to", type=str, help="Recipient phone number (+E.164 format)")
     quo_parser.add_argument("-q", "--quiet", action="store_true", help="Suppress success output")
     quo_parser.set_defaults(func=cmd_quo)
-
-    # === tunnels command group ===
-    tunnels_parser = subparsers.add_parser("tunnels", help="Manage SSH tunnels for service routing")
-    tunnels_subparsers = tunnels_parser.add_subparsers(dest="tunnels_command")
-
-    # tunnels up
-    tunnels_up = tunnels_subparsers.add_parser("up", help="Create all required tunnels")
-    tunnels_up.set_defaults(func=cmd_tunnels_up)
-
-    # tunnels down
-    tunnels_down = tunnels_subparsers.add_parser("down", help="Tear down all tunnels")
-    tunnels_down.set_defaults(func=cmd_tunnels_down)
-
-    # tunnels status
-    tunnels_status = tunnels_subparsers.add_parser("status", help="Show tunnel health")
-    tunnels_status.set_defaults(func=cmd_tunnels_status)
-
-    # tunnels check
-    tunnels_check = tunnels_subparsers.add_parser("check", help="Verify tunnels are working")
-    tunnels_check.set_defaults(func=cmd_tunnels_check)
 
     # === notify command (worker→parent) ===
     notify_cmd_parser = subparsers.add_parser("notify-parent", help="Notify parent session (worker→orchestrator)")
@@ -2580,130 +1167,6 @@ Command Categories:
     # Default listen (no subcommand) = toggle
     listen_parser.set_defaults(func=cmd_listen_toggle)
 
-    # === handoff command group ===
-    handoff_parser = subparsers.add_parser(
-        "handoff",
-        help="Shareable conversation bundles (ai-handoff.md + show-the-story.html)",
-    )
-    handoff_subparsers = handoff_parser.add_subparsers(dest="handoff_command")
-
-    handoff_init_parser = handoff_subparsers.add_parser(
-        "init",
-        help="Create a bundle dir + pre-filled ai-handoff.md template",
-    )
-    handoff_init_parser.add_argument(
-        "--title", help="Short title hint, used in the bundle slug"
-    )
-    handoff_init_parser.add_argument(
-        "--output-dir", help="Override artifacts dir (default: ~/.agentwire/artifacts)"
-    )
-    handoff_init_parser.add_argument(
-        "--json", action="store_true", help="JSON output"
-    )
-    handoff_init_parser.set_defaults(func=cmd_handoff_init)
-
-    handoff_render_parser = handoff_subparsers.add_parser(
-        "render",
-        help="Render show-the-story.html from an existing ai-handoff.md",
-    )
-    handoff_render_parser.add_argument(
-        "path", help="Bundle dir or path to ai-handoff.md"
-    )
-    handoff_render_parser.add_argument(
-        "--story", action="store_true", default=True,
-        help="Render show-the-story.html (default: on)",
-    )
-    handoff_render_parser.add_argument(
-        "--no-story", dest="story", action="store_false",
-        help="Skip HTML render (only validates the markdown)",
-    )
-    handoff_render_parser.add_argument(
-        "--json", action="store_true", help="JSON output"
-    )
-    handoff_render_parser.set_defaults(func=cmd_handoff_render)
-
-    handoff_list_parser = handoff_subparsers.add_parser(
-        "list", help="List past handoff bundles"
-    )
-    handoff_list_parser.add_argument(
-        "--json", action="store_true", help="JSON output"
-    )
-    handoff_list_parser.set_defaults(func=cmd_handoff_list)
-
-    # === roles command group ===
-    roles_parser = subparsers.add_parser(
-        "roles", help="Manage composable roles"
-    )
-    roles_subparsers = roles_parser.add_subparsers(dest="roles_command")
-
-    # roles list
-    roles_list = roles_subparsers.add_parser("list", help="List available roles")
-    roles_list.add_argument("--json", action="store_true", help="Output as JSON")
-    roles_list.set_defaults(func=cmd_roles_list)
-
-    # roles show <name>
-    roles_show = roles_subparsers.add_parser("show", help="Show role details")
-    roles_show.add_argument("name", help="Role name")
-    roles_show.add_argument("--json", action="store_true", help="Output as JSON")
-    roles_show.set_defaults(func=cmd_roles_show)
-
-    # === projects command group ===
-    projects_parser = subparsers.add_parser(
-        "projects", help="Discover and list projects"
-    )
-    projects_subparsers = projects_parser.add_subparsers(dest="projects_command")
-
-    # projects list
-    projects_list = projects_subparsers.add_parser("list", help="List discovered projects")
-    projects_list.add_argument("--machine", help="Filter by machine ID (e.g., 'local', 'mac-studio')")
-    projects_list.add_argument("--json", action="store_true", help="Output as JSON")
-    projects_list.set_defaults(func=cmd_projects_list)
-
-    # projects create
-    projects_create = projects_subparsers.add_parser(
-        "create",
-        help="Create a new local project under projects.dir with a default .agentwire.yml",
-    )
-    projects_create.add_argument("name", help="Project name (directory name under projects.dir)")
-    projects_create.add_argument(
-        "--from", dest="from_url", help="Clone from this git URL instead of creating an empty directory"
-    )
-    projects_create.add_argument(
-        "--git-init", action="store_true", help="Run 'git init' in the new project (ignored when --from is given)"
-    )
-    projects_create.add_argument("--json", action="store_true", help="Output as JSON")
-    projects_create.set_defaults(func=cmd_projects_create)
-
-    # === hooks command group ===
-    hooks_parser = subparsers.add_parser(
-        "hooks", help="Manage agentwire hook files (permission, idle handler, queue processor)"
-    )
-    hooks_subparsers = hooks_parser.add_subparsers(dest="hooks_command")
-
-    # hooks install
-    hooks_install = hooks_subparsers.add_parser(
-        "install", help="Install/refresh agentwire hook files and slash commands"
-    )
-    hooks_install.add_argument(
-        "--force", "-f", action="store_true", help="Reinstall even when already current"
-    )
-    hooks_install.add_argument(
-        "--copy", action="store_true", help="Copy files instead of symlinking"
-    )
-    hooks_install.set_defaults(func=cmd_hooks_install)
-
-    # hooks uninstall
-    hooks_uninstall = hooks_subparsers.add_parser(
-        "uninstall", help="Remove agentwire hook files and their registration"
-    )
-    hooks_uninstall.set_defaults(func=cmd_hooks_uninstall)
-
-    # hooks status
-    hooks_status = hooks_subparsers.add_parser(
-        "status", help="Check hook installation status"
-    )
-    hooks_status.set_defaults(func=cmd_hooks_status)
-
     # === generate-certs (top-level shortcut) ===
     certs_parser = subparsers.add_parser(
         "generate-certs", help="Generate SSL certificates"
@@ -2725,14 +1188,6 @@ Command Categories:
         "uninstall", help="Clear uv cache and uninstall the tool"
     )
     uninstall_parser.set_defaults(func=cmd_uninstall)
-
-    # === mcp command ===
-    mcp_parser = subparsers.add_parser(
-        "mcp",
-        help="Run MCP server for external agent integration",
-        description="Expose AgentWire as an MCP server for tools like MoltBot, Claude Desktop, etc.",
-    )
-    mcp_parser.set_defaults(func=cmd_mcp)
 
     # === scratchpad command group ===
     scratchpad_parser = subparsers.add_parser(
@@ -2817,6 +1272,7 @@ Command Categories:
     from . import channels_cli, portal_cli, tts_cli
     from . import scheduler_cli, ensure_cli
     from . import doctor_cli, history_cli, machine_cli, safety_cli
+    from . import handoff_cli, hooks_cli, mcp_cli, roles_cli, tunnels_cli
     from .council import cli as council_cli
 
     _REGISTRARS = [  # noqa: N806  # registry constant; Phase 1 of #495 appends here
@@ -2836,6 +1292,12 @@ Command Categories:
         safety_cli.register_safety_parser,
         machine_cli.register_machine_parser,
         history_cli.register_history_parser,
+        roles_cli.register_roles_parser,
+        roles_cli.register_projects_parser,
+        hooks_cli.register_hooks_parser,
+        tunnels_cli.register_tunnels_parser,
+        handoff_cli.register_handoff_parser,
+        mcp_cli.register_mcp_parser,
     ]
     for _reg in _REGISTRARS:
         _reg(subparsers)
