+++ agentwire/tunnels_cli.py
@@ -1,0 +1,253 @@
+"""CLI for SSH tunnels — ``agentwire tunnels ...``.
+
+Tunnels route portal/TTS traffic between machines. These commands create,
+tear down, and health-check the tunnels required by this machine's config.
+"""
+
+from __future__ import annotations
+
+
+def cmd_tunnels_up(args) -> int:
+    """Create all required tunnels."""
+    from .network import NetworkContext
+    from .tunnels import TunnelManager
+
+    ctx = NetworkContext.from_config()
+    manager = TunnelManager()
+    required = ctx.get_required_tunnels()
+
+    if not required:
+        print("No tunnels required for this machine's configuration.")
+        print("(All services run locally or no remote services configured)")
+        return 0
+
+    print("Creating tunnels for this machine...\n")
+
+    all_success = True
+    for i, spec in enumerate(required, 1):
+        # Get service name for display
+        service_name = _get_service_for_tunnel(ctx, spec)
+
+        print(f"[{i}/{len(required)}] {service_name} (localhost:{spec.local_port} -> {spec.remote_machine}:{spec.remote_port})")
+
+        status = manager.create_tunnel(spec, ctx)
+
+        if status.status == "up":
+            if status.error:
+                # Tunnel up but service not responding
+                print(f"      ! Tunnel created (PID {status.pid})")
+                print(f"      ! Warning: {status.error}")
+            else:
+                print(f"      + Tunnel created (PID {status.pid})")
+        else:
+            all_success = False
+            print(f"      x Failed: {status.error}")
+            _print_tunnel_help(spec, status.error)
+
+        print()
+
+    if all_success:
+        print("All tunnels up. Services should be reachable.")
+    else:
+        print("Some tunnels failed. Check errors above.")
+        return 1
+
+    return 0
+
+
+def cmd_tunnels_down(args) -> int:
+    """Tear down all tunnels."""
+    from .tunnels import TunnelManager
+
+    manager = TunnelManager()
+    count = manager.destroy_all_tunnels()
+
+    if count == 0:
+        print("No active tunnels to tear down.")
+    else:
+        print(f"Killed {count} tunnel(s).")
+
+    return 0
+
+
+def cmd_tunnels_status(args) -> int:
+    """Show tunnel health."""
+    from .network import NetworkContext
+    from .tunnels import TunnelManager
+
+    ctx = NetworkContext.from_config()
+    manager = TunnelManager()
+
+    # Get both required and active tunnels
+    required = ctx.get_required_tunnels()
+    active = manager.list_tunnels()
+
+    print("AgentWire Tunnels")
+    print("-" * 55)
+
+    if not required and not active:
+        print("\nNo tunnels configured or active.")
+        print("(All services run locally or no remote services configured)")
+        return 0
+
+    # Show required tunnels
+    for spec in required:
+        service_name = _get_service_for_tunnel(ctx, spec)
+
+        print(f"\n{service_name} (localhost:{spec.local_port} -> {spec.remote_machine}:{spec.remote_port})")
+
+        status = manager.check_tunnel(spec)
+
+        if status.status == "up":
+            print(f"  Status: + UP (PID {status.pid})")
+        elif status.status == "down":
+            print("  Status: - DOWN")
+        else:
+            print("  Status: x ERROR")
+            if status.error:
+                print(f"  Error: {status.error}")
+
+    # Show any orphaned tunnels (active but not required)
+    required_ids = {s.id for s in required}
+    orphaned = [t for t in active if t.spec.id not in required_ids]
+    if orphaned:
+        print("\n" + "-" * 55)
+        print("\nOrphaned tunnels (active but no longer required):")
+        for t in orphaned:
+            print(f"  localhost:{t.spec.local_port} -> {t.spec.remote_machine}:{t.spec.remote_port}")
+            print(f"    PID: {t.pid}, Status: {t.status}")
+
+    print("\n" + "-" * 55)
+
+    # Show next steps
+    down_tunnels = [s for s in required if manager.check_tunnel(s).status != "up"]
+    if down_tunnels:
+        print("To create missing tunnels: agentwire tunnels up")
+
+    return 0
+
+
+def cmd_tunnels_check(args) -> int:
+    """Verify tunnels are working with health checks."""
+    from .network import NetworkContext
+    from .tunnels import TunnelManager, test_service_health
+
+    ctx = NetworkContext.from_config()
+    manager = TunnelManager()
+    required = ctx.get_required_tunnels()
+
+    if not required:
+        print("No tunnels required for this machine.")
+        return 0
+
+    print("Checking tunnel health...\n")
+
+    all_healthy = True
+    for spec in required:
+        service_name = _get_service_for_tunnel(ctx, spec)
+        status = manager.check_tunnel(spec)
+
+        if status.status == "up":
+            # Also test the actual service through the tunnel
+            url = f"http://localhost:{spec.local_port}/health"
+            healthy, err = test_service_health(url, timeout=3)
+
+            if healthy:
+                print(f"+ {service_name}: healthy")
+            else:
+                print(f"! {service_name}: tunnel up but service not responding")
+                if err:
+                    print(f"  {err}")
+                all_healthy = False
+        elif status.status == "down":
+            print(f"x {service_name}: down")
+            all_healthy = False
+        else:
+            print(f"x {service_name}: error - {status.error}")
+            all_healthy = False
+
+    if all_healthy:
+        print("\nAll tunnels healthy.")
+        return 0
+    else:
+        print("\nSome tunnels need attention. Run: agentwire tunnels up")
+        return 1
+
+
+def _get_service_for_tunnel(ctx, spec) -> str:
+    """Get human-readable service name for a tunnel spec."""
+    # Check which service this tunnel is for
+    for service_name in ["portal", "tts"]:
+        service_config = getattr(ctx.config.services, service_name, None)
+        if service_config and service_config.machine == spec.remote_machine and service_config.port == spec.remote_port:
+            return f"Portal -> {service_name.upper()}" if service_name != "portal" else "Portal"
+
+    return f"Tunnel to {spec.remote_machine}"
+
+
+def _print_tunnel_help(spec, error: str) -> None:
+    """Print helpful diagnostics for tunnel errors."""
+    if not error:
+        return
+
+    error_lower = error.lower()
+
+    print("\n      Possible causes:")
+
+    if "port" in error_lower and "in use" in error_lower:
+        print("        1. Another process is using this port")
+        print("        2. A previous tunnel wasn't cleaned up")
+        print("\n      To diagnose:")
+        print(f"        lsof -i :{spec.local_port}    # Find process using port")
+        print("        agentwire tunnels down        # Clean up stale tunnels")
+
+    elif "permission denied" in error_lower:
+        print("        1. SSH key not authorized on remote machine")
+        print("        2. Wrong user configured")
+        print("\n      To fix:")
+        print(f"        ssh-copy-id {spec.remote_machine}")
+
+    elif "host key" in error_lower:
+        print("        1. Remote machine was reinstalled/changed")
+        print("        2. Possible security issue (man-in-the-middle)")
+        print("\n      If expected, fix with:")
+        print(f"        ssh-keygen -R {spec.remote_machine}")
+
+    elif "connection refused" in error_lower:
+        print("        1. SSH server not running on remote")
+        print("        2. Firewall blocking port 22")
+        print("\n      To diagnose:")
+        print(f"        ssh {spec.remote_machine} echo ok")
+
+    elif "timed out" in error_lower or "no route" in error_lower:
+        print("        1. Machine is powered off or unreachable")
+        print("        2. Network connectivity issue")
+        print("\n      To diagnose:")
+        print(f"        ping {spec.remote_machine}")
+
+    elif "not responding" in error_lower:
+        print("        1. Remote service not started")
+        print("        2. Remote service on wrong port")
+        print("\n      To diagnose:")
+        print(f"        ssh {spec.remote_machine} 'lsof -i :{spec.remote_port}'")
+
+
+def register_tunnels_parser(subparsers) -> None:
+    tunnels_parser = subparsers.add_parser("tunnels", help="Manage SSH tunnels for service routing")
+    tunnels_subparsers = tunnels_parser.add_subparsers(dest="tunnels_command")
+
+    # tunnels up
+    tunnels_up = tunnels_subparsers.add_parser("up", help="Create all required tunnels")
+    tunnels_up.set_defaults(func=cmd_tunnels_up)
+
+    # tunnels down
+    tunnels_down = tunnels_subparsers.add_parser("down", help="Tear down all tunnels")
+    tunnels_down.set_defaults(func=cmd_tunnels_down)
+
+    # tunnels status
+    tunnels_status = tunnels_subparsers.add_parser("status", help="Show tunnel health")
+    tunnels_status.set_defaults(func=cmd_tunnels_status)
+
+    # tunnels check
+    tunnels_check = tunnels_subparsers.add_parser("check", help="Verify tunnels are working")
+    tunnels_check.set_defaults(func=cmd_tunnels_check)
